# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "App for serving tender requests",
        "title": "Tender App",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
...
        "/api/tender-request-company": {
            "put": {
                "description": "Update money Tender Company by client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tender_Company"
                ],
                "summary": "Update money Tender Company",
                "parameters": [
                    {
                        "description": "Update money Tender Company",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.TenderCompany"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "update",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/tender-request-company/{id}": {
            "delete": {
                "description": "Deletes a company from a request based on the user ID and company ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tender_Company"
                ],
                "summary": "Delete company from request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/api/tenders": {
            "get": {
                "description": "Retrieves a list of tender requests based on the provided parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenders"
                ],
                "summary": "Get list of tender requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tender request status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date in the format '2006-01-02T15:04:05Z'",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date in the format '2006-01-02T15:04:05Z'",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ds.Tender"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Update Tender by admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenders"
                ],
                "summary": "Update Tender by admin",
                "parameters": [
                    {
                        "description": "updated Assembly",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.Tender"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Deletes a tender request for the given user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenders"
                ],
                "summary": "Delete tender request by user ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/api/tenders/form/{id}": {
            "put": {
                "description": "Updates the status of a transaction request with the given ID on \"завершен\"/\"отклонен\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenders"
                ],
                "summary": "Update transaction request status by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update status",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.NewStatus"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/api/tenders/{id}": {
            "get": {
                "description": "Retrieves a tender request with the given ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenders"
                ],
                "summary": "Get tender request by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tender Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.TenderDetails"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ds.Company": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "iin": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "ds.NewStatus": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "ds.Tender": {
            "type": "object",
            "properties": {
                "application_name": {
                    "type": "string"
                },
                "completion_date": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "formation_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "moderator": {
                    "$ref": "#/definitions/ds.User"
                },
                "moderator_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/ds.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "ds.TenderCompany": {
            "type": "object",
            "properties": {
                "application": {
                    "$ref": "#/definitions/ds.Company"
                },
                "cash": {
                    "type": "number"
                },
                "company_id": {
                    "type": "integer"
                },
                "tender": {
                    "$ref": "#/definitions/ds.Tender"
                },
                "tender_id": {
                    "type": "integer"
                }
            }
        },
        "ds.TenderDetails": {
            "type": "object",
            "properties": {
                "companies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Company"
                    }
                },
                "tender": {
                    "$ref": "#/definitions/ds.Tender"
                }
            }
        },
        "ds.User": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "isAdmin": {
                    "type": "boolean"
                },
                "login": {
                    "type": "string",
                    "maxLength": 64
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 8
                },
                "registrationDate": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
    }
}